+++ src/components/booksComponents/removeBook.js
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { removeBook } from '../redux/books/books';
+import { removeBook } from '../../redux/books/books';
 
 export default function RemoveBook({ bookID }) {
   const dispatch = useDispatch();
